voice_tools_v3/requirements.txt
numpy
soundfile
faster-whisper
pyyaml
